fix average result (float to int)

+++ friends-by-age.py
@@ -15,7 +15,7 @@ def parse_line(line):
 lines = sc.textFile("file:///Users/hotamul/SparkProjects/MovieRating/fakefriends.csv")
 rdd = lines.map(parse_line)
 totalsByAge = rdd.mapValues(lambda x: (x, 1)).reduceByKey(lambda x, y: (x[0] + y[0], x[1] + y[1]))
-averagesByAge = totalsByAge.mapValues(lambda x: x[0] / x[1])
+averagesByAge = totalsByAge.mapValues(lambda x: int(x[0] / x[1]))
 results = averagesByAge.collect()
 results.sort(key=lambda x: x[0])
 for result in results:
